# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}
export type BrainIntentRequestLocale = typeof BrainIntentRequestLocale[keyof typeof BrainIntentRequestLocale];


export const BrainIntentRequestLocale = {
  ar: 'ar',
  en: 'en',
} as const;

export interface BrainIntentRequest {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  text: string;
  locale: BrainIntentRequestLocale;
}

export type BrainIntentResponseIntent = typeof BrainIntentResponseIntent[keyof typeof BrainIntentResponseIntent];


export const BrainIntentResponseIntent = {
  property_search: 'property_search',
  people_requirements_search: 'people_requirements_search',
  find_matches: 'find_matches',
  unknown: 'unknown',
} as const;

export type BrainIntentResponsePropertyType = typeof BrainIntentResponsePropertyType[keyof typeof BrainIntentResponsePropertyType];


export const BrainIntentResponsePropertyType = {
  apartment: 'apartment',
  villa: 'villa',
  chalet: 'chalet',
  shop: 'shop',
  office: 'office',
  floor: 'floor',
} as const;

export type BrainIntentResponsePurpose = typeof BrainIntentResponsePurpose[keyof typeof BrainIntentResponsePurpose];


export const BrainIntentResponsePurpose = {
  rent: 'rent',
  buy: 'buy',
} as const;

export interface BrainIntentResponse {
  intent: BrainIntentResponseIntent;
  originalQuery: string;
  propertyType?: BrainIntentResponsePropertyType;
  purpose?: BrainIntentResponsePurpose;
  areaTerms?: string[];
  /** @minimum 0 */
  budgetMin?: number;
  /** @minimum 0 */
  budgetMax?: number;
  personTerms?: string[];
}

export interface BrainTranscript {
  transcript: string;
}