# lib/api-zod/src/generated/types/brainIntentRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BrainIntentRequestLocale } from './brainIntentRequestLocale';

export interface BrainIntentRequest {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  text: string;
  locale: BrainIntentRequestLocale;
}