# lib/api-zod/src/generated/types/brainIntentResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BrainIntentResponseIntent } from './brainIntentResponseIntent';
import type { BrainIntentResponsePropertyType } from './brainIntentResponsePropertyType';
import type { BrainIntentResponsePurpose } from './brainIntentResponsePurpose';

export interface BrainIntentResponse {
  intent: BrainIntentResponseIntent;
  originalQuery: string;
  propertyType?: BrainIntentResponsePropertyType;
  purpose?: BrainIntentResponsePurpose;
  areaTerms?: string[];
  /** @minimum 0 */
  budgetMin?: number;
  /** @minimum 0 */
  budgetMax?: number;
  personTerms?: string[];
}